+++ Cython/Compiler/Optimize.py
@@ -2811,6 +2811,13 @@ class OptimizeBuiltinCalls(Visitor.NodeRefCleanupMixin,
     def _handle_simple_method_object___xor__(self, node, function, args, is_unbound_method):
         return self._optimise_num_binop('Xor', node, function, args, is_unbound_method)
 
+    def _handle_simple_method_object___mod__(self, node, function, args, is_unbound_method):
+        if len(args) != 2 or not isinstance(args[1], ExprNodes.IntNode):
+            return node
+        if not args[1].has_constant_result() or not (2 <= args[1].constant_result <= 2**30):
+            return node
+        return self._optimise_num_binop('Remainder', node, function, args, is_unbound_method)
+
     def _handle_simple_method_float___add__(self, node, function, args, is_unbound_method):
         return self._optimise_num_binop('Add', node, function, args, is_unbound_method)